Use recursive navigation numbering
When using deeply nested sections, the numbering of the headings
started to get messy. See
[GitHub issue](https://github.com/timvink/mkdocs-print-site-plugin/issues/123).

This commit revamps the TOC generation, both on the sidebar and the printed
page, to always navigate the sections recursively and provide a
predictable, simple, "classic"
(as per [this
comment](https://github.com/timvink/mkdocs-print-site-plugin/pull/115#issuecomment-2346002114))
numbering scheme.

This commit also works on and expands the scope of
[this PR](https://github.com/timvink/mkdocs-print-site-plugin/pull/115),
as it has been frozen for quite some time, and its scope was slightly
more limited.

Besides, total line count has been greatly reduced, and some aspects
of the code have been simplified.

Note, however, that here we have opted to leave the inner headings
outside the TOC of the printed document, thus keeping it consistent
with the sidebar TOC, but this is different from the original approach.
It could be changed in future commits.