Definitively fix blank page issues
Introduce a slightly altered version of the old page break logic,
but with a small tweak so that the `avoid` only applies to the very
first page, and not to nested sections. Thus there is no page break
before the TOC, but there is always a page break before a section,
even when nested.